requirements.txt
streamlit == 0.89.0
numpy == 1.21.2
pandas == 1.3.3
Pillow == 8.3.2

